+++ core/src/main/scala/org/apache/spark/deploy/history/HistoryServer.scala
@@ -308,7 +308,6 @@ object HistoryServer extends Logging {
 
     val server = new HistoryServer(conf, provider, securityManager, port)
     server.bind()
-    server.attachAllHandler()
     provider.start()
 
     ShutdownHookManager.addShutdownHook { () => server.stop() }
