+++ core/src/test/scala/org/apache/spark/deploy/history/HistoryServerSuite.scala
@@ -93,7 +93,6 @@ class HistoryServerSuite extends SparkFunSuite with BeforeAndAfter with Matchers
     server = new HistoryServer(conf, provider, securityManager, 18080)
     server.initialize()
     server.bind()
-    server.attachAllHandler()
     provider.start()
     port = server.boundPort
   }
@@ -413,7 +412,6 @@ class HistoryServerSuite extends SparkFunSuite with BeforeAndAfter with Matchers
     server = new HistoryServer(myConf, provider, securityManager, 0)
     server.initialize()
     server.bind()
-    server.attachAllHandler()
     provider.start()
     val port = server.boundPort
     val metrics = server.cacheMetrics
