+++ core/src/test/scala/org/apache/spark/deploy/master/ui/MasterWebUISuite.scala
@@ -50,7 +50,6 @@ class MasterWebUISuite extends SparkFunSuite with BeforeAndAfterAll {
   override def beforeAll(): Unit = {
     super.beforeAll()
     masterWebUI.bind()
-    masterWebUI.attachAllHandler()
   }
 
   override def afterAll(): Unit = {
